// src/lib/enums/contentTypes.enum.ts
export enum ContentTypes {
  PDF = 'application/pdf',
  WORD_DOC = 'application/msword',
  WORD_DOCX = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
}

export enum FileUploadContentTypes {
  PDF = 'pdf',
  WORD_DOC = 'word_doc',
  WORD_DOCX = 'word_docx',
  IMAGE_PNG = 'image_png',
  IMAGE_JPG = 'image_jpg',
  IMAGE_JPGEG = 'image_jpeg',
  IMAGE_GIF = 'image_gif',
  IMAGE_WEBP = 'image_webp',
  VIDEO_MP4 = 'video_mp4',
  VIDEO_WEBM = 'video_webm',
  VIDEO_MKV = 'video_mkv',
  VIDEO_3GP = 'video_3gp',
}

export enum FileUploadExtensions {
  PDF = 'pdf',
  WORD_DOC = 'doc',
  WORD_DOCX = 'docx',
  IMAGE_PNG = 'png',
  IMAGE_JPG = 'jpg',
  IMAGE_JPGEG = 'jpeg',
  IMAGE_GIF = 'gif',
  IMAGE_WEBP = 'webp',
  VIDEO_MP4 = 'mp4',
  VIDEO_WEBM = 'webm',
  VIDEO_MKV = 'mkv',
  VIDEO_3GP = '3gp',
}